+++ projects/_01_ingest/cqc_api/jobs/unused_clean_cqc_location_data.py
@@ -370,6 +370,7 @@ def remove_non_social_care_locations(df: DataFrame) -> DataFrame:
     return df.where(df[CQCL.type] == LocationType.social_care_identifier)
 
 
+# Converted to polars -> Used polars functions flatten_struct_fields and impute_missing_values instead.
 def impute_historic_relationships(df: DataFrame) -> DataFrame:
     """
     Imputes historic relationships for locations in the given DataFrame.
@@ -425,6 +426,7 @@ def impute_historic_relationships(df: DataFrame) -> DataFrame:
     return df
 
 
+# Converted to polars -> Used polars functions flatten_struct_fields and impute_missing_values instead.
 def get_relationships_where_type_is_predecessor(df: DataFrame) -> DataFrame:
     """
     Filters and aggregates relationships of type 'HSCA Predecessor' for each location.
@@ -552,6 +554,7 @@ def extract_from_struct(
     return df
 
 
+# converted to polars -> projects\_01_ingest\cqc_api\fargate\utils\locations_4_clean_utils.py
 def allocate_primary_service_type(df: DataFrame) -> DataFrame:
     """
     Allocates the primary service type for each row in the DataFrame based on the descriptions in the 'imputed_gac_service_types' field.
@@ -588,6 +591,7 @@ def allocate_primary_service_type(df: DataFrame) -> DataFrame:
     return df
 
 
+# converted to polars -> projects\_01_ingest\cqc_api\fargate\utils\locations_4_clean_utils.py
 def realign_carehome_column_with_primary_service(df: DataFrame) -> DataFrame:
     """
     Allocates the location as a care_home if primary_service_type is a care home.
@@ -619,6 +623,7 @@ def realign_carehome_column_with_primary_service(df: DataFrame) -> DataFrame:
     return df
 
 
+# converted to polars -> projects\_01_ingest\cqc_api\fargate\utils\locations_4_clean_utils.py
 def add_related_location_column(df: DataFrame) -> DataFrame:
     """
     Adds a column which flags whether the location was related to a previous location or not